pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cmdbox"
dynamic = ["version"] 
description = "cmdbox: It is a command line application with a plugin mechanism."
readme = "README.md"
authors = [
    { name = "hamacom2004jp", email = "hamacom2004jp@gmail.com" },
]
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = [
    "audit",
    "cli",
    "excel",
    "fastapi",
    "mcp",
    "redis",
    "restapi",
    "web"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Japanese",
    "Programming Language :: Python",
    "Topic :: Utilities"
]
dependencies = [
    "argcomplete",
    "aiosqlite",
    "async-timeout",
    "asyncpg",
    "cffi",
    "cryptography",
    "fastapi",
    "debugpy",
    "gevent",
    "gunicorn",
    "itsdangerous",
    "numpy",
    "openpyxl",
    "Pillow",
    "plyer",
    "psycopg[binary]",
    "pyjwt",
    "pystray",
    "python-multipart",
    "pycryptodome",
    "pyyaml",
    "questionary",
    "redis",
    "requests",
    "rich",
    "six",
    "tabulate",
    "uvicorn",
    "wsproto",
    "wheel",
]

[project.optional-dependencies]
agent_mcp = [
    "pydantic>=2.0.0,<3.0.0",
    "fastmcp>=2.14.4",
    "mcp>=1.23.0,<2.0.0",
]
agent_adk = [
    "google-adk>=1.23.0",
    "a2a-sdk>=0.3.4,<0.4.0",
    "litellm>=1.75.5,<1.80.17",
]
agent_mem = [
    "sentence-transformers",
    "sqlite_vec",
    "transformers>=4.48.0",
    "torch",
]
saml = [
    "python3-saml",
    "xmlsec==1.3.13",
]

[project.urls]
Homepage = "https://github.com/hamacom2004jp/cmdbox"
Documentation = "https://hamacom2004jp.github.io/cmdbox/index.html"
Repository = "https://github.com/hamacom2004jp/cmdbox"

[project.scripts]
cmdbox = "cmdbox.app.app:main"

# ----------------------------------------------------------------------
# setuptools固有の設定 (setup.cfgに相当する内容をsetuptoolsテーブルで記述)
# ----------------------------------------------------------------------

[tool.setuptools]
packages = [
    "cmdbox",
    "cmdbox.app",
    "cmdbox.app.auth",
    "cmdbox.app.commons",
    "cmdbox.app.features.cli",
    "cmdbox.app.features.cli.agent",
    "cmdbox.app.features.web",
    "cmdbox.extensions"
]

[tool.setuptools.package-data]
# setup()のpackage_dataとexclude_package_dataに相当
cmdbox = [
    "*.yml",
    "extensions/**",
    "extensions/sample_project/.vscode/**",
    "docker/**",
    "licenses/*",
    "tools/datas/**",
    "web/**",
]

[tool.setuptools.exclude-package-data]
cmdbox = [
    "extensions/data/*.json",
    "extensions/data/*/*.jpg",
    "extensions/data/*/*.svg",
]
